+++ src/Mscc.GenerativeAI.Web/CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Fixed
 
+## 1.8.2
+
+### Changed
+
+- bump version
+
 ## 1.8.1
 
 ### Changed